requirements.txt
astroid==2.4.0
autopep8==1.4.4
certifi==2020.4.5.1
chardet==3.0.4
click==7.1.2
colorama==0.4.3
flash==1.0.3
Flask==1.1.2
idna==2.9
isort==4.3.21
itsdangerous==1.1.0
Jinja2==2.11.2
lazy-object-proxy==1.4.3
MarkupSafe==1.1.1
mccabe==0.6.1
numpy==1.18.4
opencv-python==4.2.0.34
Pillow==7.1.2
pycodestyle==2.5.0
pylint==2.5.0
pytesseract==0.3.4
requests==2.23.0
six==1.14.0
toml==0.10.0
typed-ast==1.4.1
urllib3==1.25.9
Werkzeug==1.0.1
wincertstore==0.2
wrapt==1.11.2
